==> ner_models/utils/4_quantitative_anaylis_iob_tags.ipynb <==
import pandas as pd
import numpy as np

## Tokens Dataset

data = pd.read_csv("laudos_1_963_iob_wt_abd.csv", encoding = 'utf-8', index_col=0)

tags_count = data.tag.value_counts()

tags_count

total_tokens = tags_count.sum()

total_tokens

ntr_o = (tags_count[0]/total_tokens)*100
ntr_b_achado = (tags_count[4]/total_tokens)*100
ntr_i_achado = (tags_count[6]/total_tokens)*100
ntr_b_atenucao = (tags_count[11]/total_tokens)*100
ntr_i_atenuacao = (tags_count[8]/total_tokens)*100
ntr_b_bordas = (tags_count[12]/total_tokens)*100
ntr_i_bordas = (tags_count[10]/total_tokens)*100
ntr_b_calcificacao = (tags_count[7]/total_tokens)*100
ntr_i_calcificacao = (tags_count[9]/total_tokens)*100
ntr_b_localizacao = (tags_count[3]/total_tokens)*100
ntr_i_localizacao = (tags_count[1]/total_tokens)*100
ntr_b_tamanho = (tags_count[5]/total_tokens)*100
ntr_i_tamanho = (tags_count[2]/total_tokens)*100

print(round(ntr_b_achado, 2))
print(round(ntr_i_achado, 2))
print()
print(round(ntr_b_atenucao, 2))
print(round(ntr_i_atenuacao, 2))
print()
print(round(ntr_b_bordas, 2))
print(round(ntr_i_bordas, 2))
print()
print(round(ntr_b_calcificacao, 2))
print(round(ntr_i_calcificacao, 2))
print()
print(round(ntr_b_localizacao, 2))
print(round(ntr_i_localizacao, 2))
print()
print(round(ntr_b_tamanho, 2))
print(round(ntr_i_tamanho, 2))
print()
print(round(ntr_o, 2))

# Unique Tokens Dataset

data_b_achado = data[data.tag == 'B-Achado']
data_i_achado = data[data.tag == 'I-Achado']
data_b_atenuacao = data[data.tag == 'B-Atenuação']
data_i_atenuacao = data[data.tag == 'I-Atenuação']
data_b_bordas = data[data.tag == 'B-Bordas']
data_i_bordas = data[data.tag == 'I-Bordas']
data_b_calcificacao = data[data.tag == 'B-Calcificação']
data_i_calcificacao = data[data.tag == 'I-Calcificação']
data_b_localizacao = data[data.tag == 'B-Localização']
data_i_localizacao = data[data.tag == 'I-Localização']
data_b_tamanho = data[data.tag == 'B-Tamanho']
data_i_tamanho = data[data.tag == 'I-Tamanho']
data_o = data[data.tag == 'O']

# Unique Tokens Entity Dataset

ndt_b_achado = len(data_b_achado.word.unique())
ndt_i_achado = len(data_i_achado.word.unique())

ndt_b_atenuacao = len(data_b_atenuacao.word.unique())
ndt_i_atenuacao = len(data_i_atenuacao.word.unique())

ndt_b_bordas = len(data_b_bordas.word.unique())
ndt_i_bordas = len(data_i_bordas.word.unique())

ndt_b_calcificacao = len(data_b_calcificacao.word.unique())
ndt_i_calcificacao = len(data_i_calcificacao.word.unique())

ndt_b_localizacao = len(data_b_localizacao.word.unique())
ndt_i_localizacao = len(data_i_localizacao.word.unique())

ndt_b_tamanho = len(data_b_tamanho.word.unique())
ndt_i_tamanho = len(data_i_tamanho.word.unique())

ndt_o = len(data_o.word.unique())

print(round(ndt_b_achado, 2))
print(round(ndt_i_achado, 2))
print()
print(round(ndt_b_atenuacao, 2))
print(round(ndt_i_atenuacao, 2))
print()
print(round(ndt_b_bordas, 2))
print(round(ndt_i_bordas, 2))
print()
print(round(ndt_b_calcificacao, 2))
print(round(ndt_i_calcificacao, 2))
print()
print(round(ndt_b_localizacao, 2))
print(round(ndt_i_localizacao, 2))
print()
print(round(ndt_b_tamanho, 2))
print(round(ndt_i_tamanho, 2))
print()
print(round(ndt_o, 2))

ndt_unique = (ndt_b_achado + ndt_i_achado
             + ndt_b_atenuacao + ndt_i_atenuacao
             + ndt_b_bordas + ndt_i_bordas
             + ndt_b_calcificacao + ndt_i_calcificacao
             + ndt_b_localizacao + ndt_i_localizacao
             + ndt_b_tamanho + ndt_i_tamanho
             + ndt_o)

print(ndt_unique)

ndtr_b_achado = (ndt_b_achado/ndt_unique)*100
ndtr_i_achado = (ndt_i_achado/ndt_unique)*100

ndtr_b_atenuacao = (ndt_b_atenuacao/ndt_unique)*100
ndtr_i_atenuacao = (ndt_i_atenuacao/ndt_unique)*100

ndtr_b_bordas = (ndt_b_bordas/ndt_unique)*100
ndtr_i_bordas = (ndt_i_bordas/ndt_unique)*100

ndtr_b_calcificacao = (ndt_b_calcificacao/ndt_unique)*100
ndtr_i_calcificacao = (ndt_i_calcificacao/ndt_unique)*100

ndtr_b_localizacao = (ndt_b_localizacao/ndt_unique)*100
ndtr_i_localizacao = (ndt_i_localizacao/ndt_unique)*100

ndtr_b_tamanho = (ndt_b_tamanho/ndt_unique)*100
ndtr_i_tamanho = (ndt_i_tamanho/ndt_unique)*100

ndtr_o = (ndt_o/ndt_unique)*100

print(round(ndtr_b_achado, 2))
print(round(ndtr_i_achado, 2))
print()
print(round(ndtr_b_atenuacao, 2))
print(round(ndtr_i_atenuacao, 2))
print()
print(round(ndtr_b_bordas, 2))
print(round(ndtr_i_bordas, 2))
print()
print(round(ndtr_b_calcificacao, 2))
print(round(ndtr_i_calcificacao, 2))
print()
print(round(ndtr_b_localizacao, 2))
print(round(ndtr_i_localizacao, 2))
print()
print(round(ndtr_b_tamanho, 2))
print(round(ndtr_i_tamanho, 2))
print()
print(round(ndtr_o, 2))

# Tokens Test Dataset

import os

print(os.getcwd())

cwd = "C:\\Users\\tarci\\Desktop\\modelos_mestrado\\bilstm_antigo\\results\\"

data_test = pd.read_csv(cwd+"result_df_model_20.csv",  encoding = 'utf-8', index_col=0)

data_test.head()

tags_count_test = data_test.tag.value_counts()

tags_count_test

total_tokens_test = tags_count_test.sum()

total_tokens_test

ntr_o_test = (tags_count_test[0]/total_tokens_test)*100
ntr_b_achado_test = (tags_count_test[4]/total_tokens_test)*100
ntr_i_achado_test = (tags_count_test[6]/total_tokens_test)*100
ntr_b_atenucao_test = (tags_count_test[11]/total_tokens_test)*100
ntr_i_atenuacao_test = (tags_count_test[8]/total_tokens_test)*100
ntr_b_bordas_test = (tags_count_test[12]/total_tokens_test)*100
ntr_i_bordas_test = (tags_count_test[10]/total_tokens_test)*100
ntr_b_calcificacao_test = (tags_count_test[7]/total_tokens_test)*100
ntr_i_calcificacao_test = (tags_count_test[9]/total_tokens_test)*100
ntr_b_localizacao_test = (tags_count_test[3]/total_tokens_test)*100
ntr_i_localizacao_test = (tags_count_test[1]/total_tokens_test)*100
ntr_b_tamanho_test = (tags_count_test[5]/total_tokens_test)*100
ntr_i_tamanho_test = (tags_count_test[2]/total_tokens_test)*100

print(round(ntr_b_achado_test, 3))
print(round(ntr_i_achado_test, 3))
print()
print(round(ntr_b_atenucao_test, 3))
print(round(ntr_i_atenuacao_test, 3))
print()
print(round(ntr_b_bordas_test, 3))
print(round(ntr_i_bordas_test, 3))
print()
print(round(ntr_b_calcificacao_test, 3))
print(round(ntr_i_calcificacao_test, 3))
print()
print(round(ntr_b_localizacao_test, 3))
print(round(ntr_i_localizacao_test, 3))
print()
print(round(ntr_b_tamanho_test, 3))
print(round(ntr_i_tamanho_test, 3))
print()
print(round(ntr_o_test, 3))

# Unique Tokens Test Dataset

data_b_achado_test = data_test[data_test.tag == 'B-Achado']
data_i_achado_test = data_test[data_test.tag == 'I-Achado']
data_b_atenuacao_test = data_test[data_test.tag == 'B-Atenuação']
data_i_atenuacao_test = data_test[data_test.tag == 'I-Atenuação']
data_b_bordas_test = data_test[data_test.tag == 'B-Bordas']
data_i_bordas_test = data_test[data_test.tag == 'I-Bordas']
data_b_calcificacao_test = data_test[data_test.tag == 'B-Calcificação']
data_i_calcificacao_test = data_test[data_test.tag == 'I-Calcificação']
data_b_localizacao_test = data_test[data_test.tag == 'B-Localização']
data_i_localizacao_test = data_test[data_test.tag == 'I-Localização']
data_b_tamanho_test = data_test[data_test.tag == 'B-Tamanho']
data_i_tamanho_test = data_test[data_test.tag == 'I-Tamanho']
data_o_test = data_test[data_test.tag == 'O']

# Unique Tokens Entity Test Dataset

ndt_b_achado_test = len(data_b_achado_test.word.unique())
ndt_i_achado_test = len(data_i_achado_test.word.unique())

ndt_b_atenuacao_test = len(data_b_atenuacao_test.word.unique())
ndt_i_atenuacao_test = len(data_i_atenuacao_test.word.unique())

ndt_b_bordas_test = len(data_b_bordas_test.word.unique())
ndt_i_bordas_test = len(data_i_bordas_test.word.unique())

ndt_b_calcificacao_test = len(data_b_calcificacao_test.word.unique())
ndt_i_calcificacao_test = len(data_i_calcificacao_test.word.unique())

ndt_b_localizacao_test = len(data_b_localizacao_test.word.unique())
ndt_i_localizacao_test = len(data_i_localizacao_test.word.unique())

ndt_b_tamanho_test = len(data_b_tamanho_test.word.unique())
ndt_i_tamanho_test = len(data_i_tamanho_test.word.unique())

ndt_o_test = len(data_o_test.word.unique())

print(round(ndt_b_achado_test, 3))
print(round(ndt_i_achado_test, 3))
print()
print(round(ndt_b_atenuacao_test, 3))
print(round(ndt_i_atenuacao_test, 3))
print()
print(round(ndt_b_bordas_test, 3))
print(round(ndt_i_bordas_test, 3))
print()
print(round(ndt_b_calcificacao_test, 3))
print(round(ndt_i_calcificacao_test, 3))
print()
print(round(ndt_b_localizacao_test, 3))
print(round(ndt_i_localizacao_test, 3))
print()
print(round(ndt_b_tamanho_test, 3))
print(round(ndt_i_tamanho_test, 3))
print()
print(round(ndt_o_test, 3))

ndt_unique_test = (ndt_b_achado_test + ndt_i_achado_test
             + ndt_b_atenuacao_test + ndt_i_atenuacao_test
             + ndt_b_bordas_test + ndt_i_bordas_test
             + ndt_b_calcificacao_test + ndt_i_calcificacao_test
             + ndt_b_localizacao_test + ndt_i_localizacao_test
             + ndt_b_tamanho_test + ndt_i_tamanho_test
             + ndt_o_test)

print(ndt_unique_test)

ndtr_b_achado_test = (ndt_b_achado_test/ndt_unique_test)*100
ndtr_i_achado_test = (ndt_i_achado_test/ndt_unique_test)*100

ndtr_b_atenuacao_test = (ndt_b_atenuacao_test/ndt_unique_test)*100
ndtr_i_atenuacao_test = (ndt_i_atenuacao_test/ndt_unique_test)*100

ndtr_b_bordas_test = (ndt_b_bordas_test/ndt_unique_test)*100
ndtr_i_bordas_test = (ndt_i_bordas_test/ndt_unique_test)*100

ndtr_b_calcificacao_test = (ndt_b_calcificacao_test/ndt_unique_test)*100
ndtr_i_calcificacao_test = (ndt_i_calcificacao_test/ndt_unique_test)*100

ndtr_b_localizacao_test = (ndt_b_localizacao_test/ndt_unique_test)*100
ndtr_i_localizacao_test = (ndt_i_localizacao_test/ndt_unique_test)*100

ndtr_b_tamanho_test = (ndt_b_tamanho_test/ndt_unique_test)*100
ndtr_i_tamanho_test = (ndt_i_tamanho_test/ndt_unique_test)*100

ndtr_o_test = (ndt_o_test/ndt_unique_test)*100

print(round(ndtr_b_achado_test, 3))
print(round(ndtr_i_achado_test, 3))
print()
print(round(ndtr_b_atenuacao_test, 3))
print(round(ndtr_i_atenuacao_test, 3))
print()
print(round(ndtr_b_bordas_test, 3))
print(round(ndtr_i_bordas_test, 3))
print()
print(round(ndtr_b_calcificacao_test, 3))
print(round(ndtr_i_calcificacao_test, 3))
print()
print(round(ndtr_b_localizacao_test, 3))
print(round(ndtr_i_localizacao_test, 3))
print()
print(round(ndtr_b_tamanho_test, 3))
print(round(ndtr_i_tamanho_test, 3))
print()
print(round(ndtr_o_test, 3))

# Tokens Training Dataset

data_train = pd.read_csv("train_df.csv", encoding = 'utf-8', index_col=0)

tags_count_train = data_train.tag.value_counts()

tags_count_train

total_tokens_train = tags_count_train.sum()

total_tokens_train

ntr_o_train = (tags_count_train[0]/total_tokens_train)*100
ntr_b_achado_train = (tags_count_train[4]/total_tokens_train)*100
ntr_i_achado_train = (tags_count_train[6]/total_tokens_train)*100
ntr_b_atenucao_train = (tags_count_train[11]/total_tokens_train)*100
ntr_i_atenuacao_train = (tags_count_train[8]/total_tokens_train)*100
ntr_b_bordas_train = (tags_count_train[12]/total_tokens_train)*100
ntr_i_bordas_train = (tags_count_train[10]/total_tokens_train)*100
ntr_b_calcificacao_train = (tags_count_train[7]/total_tokens_train)*100
ntr_i_calcificacao_train = (tags_count_train[9]/total_tokens_train)*100
ntr_b_localizacao_train = (tags_count_train[3]/total_tokens_train)*100
ntr_i_localizacao_train = (tags_count_train[1]/total_tokens_train)*100
ntr_b_tamanho_train = (tags_count_train[5]/total_tokens_train)*100
ntr_i_tamanho_train = (tags_count_train[2]/total_tokens_train)*100

print(round(ntr_b_achado_train, 3))
print(round(ntr_i_achado_train, 3))
print()
print(round(ntr_b_atenucao_train, 3))
print(round(ntr_i_atenuacao_train, 3))
print()
print(round(ntr_b_bordas_train, 3))
print(round(ntr_i_bordas_train, 3))
print()
print(round(ntr_b_calcificacao_train, 3))
print(round(ntr_i_calcificacao_train, 3))
print()
print(round(ntr_b_localizacao_train, 3))
print(round(ntr_i_localizacao_train, 3))
print()
print(round(ntr_b_tamanho_train, 3))
print(round(ntr_i_tamanho_train, 3))
print()
print(round(ntr_o_train, 3))

# Unique Tokens Training Dataset

data_b_achado_train = data_train[data_train.tag == 'B-Achado']
data_i_achado_train = data_train[data_train.tag == 'I-Achado']
data_b_atenuacao_train = data_train[data_train.tag == 'B-Atenuação']
data_i_atenuacao_train = data_train[data_train.tag == 'I-Atenuação']
data_b_bordas_train = data_train[data_train.tag == 'B-Bordas']
data_i_bordas_train = data_train[data_train.tag == 'I-Bordas']
data_b_calcificacao_train = data_train[data_train.tag == 'B-Calcificação']
data_i_calcificacao_train = data_train[data_train.tag == 'I-Calcificação']
data_b_localizacao_train = data_train[data_train.tag == 'B-Localização']
data_i_localizacao_train = data_train[data_train.tag == 'I-Localização']
data_b_tamanho_train = data_train[data_train.tag == 'B-Tamanho']
data_i_tamanho_train = data_train[data_train.tag == 'I-Tamanho']
data_o_train = data_train[data_train.tag == 'O']

ndt_b_achado_train = len(data_b_achado_train.word.unique())
ndt_i_achado_train = len(data_i_achado_train.word.unique())

ndt_b_atenuacao_train = len(data_b_atenuacao_train.word.unique())
ndt_i_atenuacao_train = len(data_i_atenuacao_train.word.unique())

ndt_b_bordas_train = len(data_b_bordas_train.word.unique())
ndt_i_bordas_train = len(data_i_bordas_train.word.unique())

ndt_b_calcificacao_train = len(data_b_calcificacao_train.word.unique())
ndt_i_calcificacao_train = len(data_i_calcificacao_train.word.unique())

ndt_b_localizacao_train = len(data_b_localizacao_train.word.unique())
ndt_i_localizacao_train = len(data_i_localizacao_train.word.unique())

ndt_b_tamanho_train = len(data_b_tamanho_train.word.unique())
ndt_i_tamanho_train = len(data_i_tamanho_train.word.unique())

ndt_o_train = len(data_o_train.word.unique())

print(round(ndt_b_achado_train, 3))
print(round(ndt_i_achado_train, 3))
print()
print(round(ndt_b_atenuacao_train, 3))
print(round(ndt_i_atenuacao_train, 3))
print()
print(round(ndt_b_bordas_train, 3))
print(round(ndt_i_bordas_train, 3))
print()
print(round(ndt_b_calcificacao_train, 3))
print(round(ndt_i_calcificacao_train, 3))
print()
print(round(ndt_b_localizacao_train, 3))
print(round(ndt_i_localizacao_train, 3))
print()
print(round(ndt_b_tamanho_train, 3))
print(round(ndt_i_tamanho_train, 3))
print()
print(round(ndt_o_train, 3))

ndt_unique_train = (ndt_b_achado_train + ndt_i_achado_train
             + ndt_b_atenuacao_train + ndt_i_atenuacao_train
             + ndt_b_bordas_train + ndt_i_bordas_train
             + ndt_b_calcificacao_train + ndt_i_calcificacao_train
             + ndt_b_localizacao_train + ndt_i_localizacao_train
             + ndt_b_tamanho_train + ndt_i_tamanho_train
             + ndt_o_train)

ndt_unique_train

ndtr_b_achado_train = (ndt_b_achado_train/ndt_unique_train)*100
ndtr_i_achado_train = (ndt_i_achado_train/ndt_unique_train)*100

ndtr_b_atenuacao_train = (ndt_b_atenuacao_train/ndt_unique_train)*100
ndtr_i_atenuacao_train = (ndt_i_atenuacao_train/ndt_unique_train)*100

ndtr_b_bordas_train = (ndt_b_bordas_train/ndt_unique_train)*100
ndtr_i_bordas_train = (ndt_i_bordas_train/ndt_unique_train)*100

ndtr_b_calcificacao_train = (ndt_b_calcificacao_train/ndt_unique_train)*100
ndtr_i_calcificacao_train = (ndt_i_calcificacao_train/ndt_unique_train)*100

ndtr_b_localizacao_train = (ndt_b_localizacao_train/ndt_unique_train)*100
ndtr_i_localizacao_train = (ndt_i_localizacao_train/ndt_unique_train)*100

ndtr_b_tamanho_train = (ndt_b_tamanho_train/ndt_unique_train)*100
ndtr_i_tamanho_train = (ndt_i_tamanho_train/ndt_unique_train)*100

ndtr_o_train = (ndt_o_train/ndt_unique_train)*100

print(round(ndtr_b_achado_train, 3))
print(round(ndtr_i_achado_train, 3))
print()
print(round(ndtr_b_atenuacao_train, 3))
print(round(ndtr_i_atenuacao_train, 3))
print()
print(round(ndtr_b_bordas_train, 3))
print(round(ndtr_i_bordas_train, 3))
print()
print(round(ndtr_b_calcificacao_train, 3))
print(round(ndtr_i_calcificacao_train, 3))
print()
print(round(ndtr_b_localizacao_train, 3))
print(round(ndtr_i_localizacao_train, 3))
print()
print(round(ndtr_b_tamanho_train, 3))
print(round(ndtr_i_tamanho_train, 3))
print()
print(round(ndtr_o_train, 3))
